製造指摘二次修正

+++ backend/src/main/java/com/example/stock/constants/BusinessConstants.java
@@ -7,9 +7,9 @@ public final class BusinessConstants {
 
     public static final Long LOGIN_USER_ID = 1L;
 
-    public static final long MAX_FAVORITE_COUNT = 20L;
-    public static final long MAX_ADMIN_STOCK_COUNT = 100L;
-    public static final long MAX_ADMIN_USER_DISPLAY_COUNT = 100L;
+    public static final int MAX_FAVORITE_COUNT = 20;
+    public static final int MAX_ADMIN_STOCK_COUNT = 100;
+    public static final int MAX_ADMIN_USER_DISPLAY_COUNT = 100;
 
     public static final int WEEK_CHART_DAYS = 7;
     public static final int MONTH_CHART_DAYS = 30;
