+++ backend/src/main/java/com/example/stock/dto/AdminStockListResponse.java
@@ -8,7 +8,6 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 public class AdminStockListResponse {
-
-    private long totalCount;
+    private int totalCount;
     private List<AdminStockItemResponse> items;
 }
