+++ backend/src/main/java/com/example/stock/dto/StockListResponse.java
@@ -8,11 +8,11 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 public class StockListResponse {
-    private long totalCount;
+    private int totalCount;
     private int page;
     private int size;
     private int totalPages;
-    private long currentFavoriteCount;
-    private long maxFavoriteCount;
+    private int currentFavoriteCount;
+    private int maxFavoriteCount;
     private List<StockListItemResponse> items;
 }
